blockdata: move TestBlock4423_KnownFeeValues from explorer to blockdata
