+++ cmd/dcrdata/internal/explorer/explorer_test.go
@@ -15,7 +15,6 @@ import (
 	explorerTypes "github.com/monetarium/monetarium-explorer/explorer/types"
 	"github.com/monetarium/monetarium-node/chaincfg"
 	"github.com/monetarium/monetarium-node/chaincfg/chainhash"
-	"github.com/monetarium/monetarium-node/dcrutil"
 	chainjson "github.com/monetarium/monetarium-node/rpc/jsonrpc/types"
 	"github.com/monetarium/monetarium-node/wire"
 )
@@ -585,24 +584,4 @@ func TestStore_PropagatesMiningFeeAtoms(t *testing.T) {
 	}
 }
 
-// Block 4423 known totals (from real block analysis):
-//   - Regular tx fees (non-coinbase): 31,130 atoms
-//   - Ticket purchase fees: 21,140 atoms
-//   - Total VAR fees: 31,130 + 21,140 = 52,270 atoms
-//   - dcrutil.Amount(52_270).ToCoin() = 0.00052270 VAR
-//   - Miner fee (coinbase P2PKH - PoW subsidy): 26,135 atoms
-func TestBlock4423_KnownFeeValues(t *testing.T) {
-	const (
-		regFees      = int64(31_130)
-		ticketFees   = int64(21_140)
-		totalVARFees = int64(52_270)
-		totalVARCoin = 0.00052270
-	)
-	if regFees+ticketFees != totalVARFees {
-		t.Fatal("broken test: reg + ticket != total")
-	}
-	coin := dcrutil.Amount(totalVARFees).ToCoin()
-	if coin != totalVARCoin {
-		t.Errorf("dcrutil.Amount(%d).ToCoin() = %.8f, want %.8f", totalVARFees, coin, totalVARCoin)
-	}
-}
+
